+++ Sources/Entities/ImageByNameInfo.swift
@@ -12,7 +12,7 @@ public struct ImageByNameInfo: Codable, Hashable {
     /// Gets or sets the context.
     public var context: String?
     /// Gets or sets the length of the file.
-    public var fileLength: Int64?
+    public var fileLength: Int?
     /// Gets or sets the format.
     public var format: String?
     /// Gets or sets the name.
@@ -20,7 +20,7 @@ public struct ImageByNameInfo: Codable, Hashable {
     /// Gets or sets the theme.
     public var theme: String?
 
-    public init(context: String? = nil, fileLength: Int64? = nil, format: String? = nil, name: String? = nil, theme: String? = nil) {
+    public init(context: String? = nil, fileLength: Int? = nil, format: String? = nil, name: String? = nil, theme: String? = nil) {
         self.context = context
         self.fileLength = fileLength
         self.format = format
@@ -31,7 +31,7 @@ public struct ImageByNameInfo: Codable, Hashable {
     public init(from decoder: Decoder) throws {
         let values = try decoder.container(keyedBy: StringCodingKey.self)
         self.context = try values.decodeIfPresent(String.self, forKey: "Context")
-        self.fileLength = try values.decodeIfPresent(Int64.self, forKey: "FileLength")
+        self.fileLength = try values.decodeIfPresent(Int.self, forKey: "FileLength")
         self.format = try values.decodeIfPresent(String.self, forKey: "Format")
         self.name = try values.decodeIfPresent(String.self, forKey: "Name")
         self.theme = try values.decodeIfPresent(String.self, forKey: "Theme")
