+++ Sources/Entities/SeekRequestDto.swift
@@ -11,15 +11,15 @@ import Foundation
 /// Class SeekRequestDto.
 public struct SeekRequestDto: Codable, Hashable {
     /// Gets or sets the position ticks.
-    public var positionTicks: Int64?
+    public var positionTicks: Int?
 
-    public init(positionTicks: Int64? = nil) {
+    public init(positionTicks: Int? = nil) {
         self.positionTicks = positionTicks
     }
 
     public init(from decoder: Decoder) throws {
         let values = try decoder.container(keyedBy: StringCodingKey.self)
-        self.positionTicks = try values.decodeIfPresent(Int64.self, forKey: "PositionTicks")
+        self.positionTicks = try values.decodeIfPresent(Int.self, forKey: "PositionTicks")
     }
 
     public func encode(to encoder: Encoder) throws {
